+++ src/net/JobResult.h
@@ -36,7 +36,7 @@ class JobResult
 {
 public:
     inline JobResult() : poolId(0), diff(0), nonce(0) {}
-    inline JobResult(int poolId, const xmrig::Id &jobId, uint32_t nonce, const uint8_t *result, uint32_t diff) :
+    inline JobResult(int poolId, const xmrig::Id &jobId, uint64_t nonce, const uint8_t *result, uint32_t diff) :
         poolId(poolId),
         diff(diff),
         nonce(nonce),
@@ -72,7 +72,7 @@ public:
 
     int poolId;
     uint32_t diff;
-    uint32_t nonce;
+    uint64_t nonce;
     uint8_t result[32];
     xmrig::Id jobId;
 };
